SGF-283 - Removed obsolete repository declarations.
To make sure the release automation has a single repository declaration to work with I removed the obsolete release and milestone repository declarations. The snapshots contains the milestone repo, which in turn contains the release one, so that only one of them has to be declared depending on which artifacts you want to see.