SGF-289 - Enumeration restrictions (xsd:enumeration) should be avoided in the XML schema.
Refactoring SDG enum types.  Renamed the EvictionType enum to EvictionPolicyType to reflect that the enum represents an Eviction 'policy' encompassing both Eviction algorith along with the monitored resource triggering Eviction.  Renamed the EvictionTypeConverter to EvictionPolicyConverter to more accurately reflect it's intent.  Renamed IndexMaintenanceType enum to IndexMaintenancePolicyType to more accurately describe the purpose of the index maintenance setting reflected as a policy.  Renamed IndexMaintenanceTypeConverter to IndexMaintenancePolicyConverter to more accurately reflect it's intent.